requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.2
Flask-WTF==1.2.1
gunicorn==23.0.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
packaging==24.1
pandas==2.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
setuptools==69.1.1
six==1.16.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.1
wheel==0.42.0
WTForms==3.1.2
resend>=0.3